requirements.txt
streamlit>=1.24.0
yt-dlp>=2023.11.16
openai-whisper==20231117
deep-translator>=1.11.4
python-dotenv>=1.0.0
youtube_transcript_api>=0.6.1
ffmpeg-python>=0.2.0
torch>=2.0.0
numpy>=1.22.0
tqdm>=4.65.0
setuptools>=65.5.1
requests>=2.31.0
urllib3>=2.0.0
google-api-python-client>=2.0.0
tenacity>=8.0.0
ffmpeg-python>=0.2.0